Reorganize folder structure - Add Mapzen.Unity namespace for Unity related script - Adjust main camera
