+++ Assets/Editor/FilterStyleEditor.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Collections.Generic;
 using Mapzen;
+using Mapzen.Unity;
 using UnityEditor;
 using UnityEngine;
