+++ Assets/Mapzen/SceneGroup.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Mapzen.Unity;
 
 namespace Mapzen
 {
