+++ Assets/Mapzen/Unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 0f7e005d222b74109bc7e27644b85d32
+folderAsset: yes
+timeCreated: 1504292054
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
