+++ Assets/Mapzen/Unity/AsyncWorker.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9be69fd3ec6204c7796f99c01ffb7534
-timeCreated: 1495055370
+guid: 7947694d47dcf494e876d6b30e227eb0
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
