+++ Assets/Mapzen/Unity/Earcut.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3dc0caad831fe4a66a931e540b657d76
-timeCreated: 1495055370
+guid: f263b0328840648838565a6de9295532
+timeCreated: 1504292877
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
