+++ Assets/Mapzen/Unity/FeatureStyle.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ba68eafcdbbee41608cd62fe4d5af22b
-timeCreated: 1495649081
+guid: bdaa043e324e04b91a4dcefec1ebb409
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
