+++ Assets/Mapzen/Unity/GameObjectOptions.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 081f0ebf1ffff4cd8a074685f33a6c8d
-timeCreated: 1498675503
+guid: 5d7cee0bbce184326869839a84d891ad
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
