+++ Assets/Mapzen/Unity/IO.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0eb0cbbdcd42b4a18805177b7dc9da97
-timeCreated: 1500668439
+guid: 9e14c57d974e840b8a79ebd30bab9b93
+timeCreated: 1504292171
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
