+++ Assets/Mapzen/Unity/MeshData.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ee2fc3e9878e94f649090a2e12d9c068
-timeCreated: 1501876542
+guid: bf2770cf1171b4486988f0a6cfb0763f
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
