+++ Assets/Mapzen/Unity/PolygonBuilder.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 812be603472a14c0999919d732c56ce7
-timeCreated: 1504207854
+guid: 4e59cd67dccdb4578bca5785208849c8
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
