+++ Assets/Mapzen/Unity/PolylineBuilder.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4283c2718eeb8404ebf83d384f29fde9
-timeCreated: 1496871192
+guid: 3c09a2ce335e646aa8f114093e019d88
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
