+++ Assets/Mapzen/Unity/TileTask.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: aa1567adc469948978f94b9eebb00927
-timeCreated: 1498675503
+guid: 5392c1502bca34312b8aa62ef1534924
+timeCreated: 1504291857
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
