+++ Assets/Scenes/DefaultStyle.asset
@@ -8,7 +8,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0eb0cbbdcd42b4a18805177b7dc9da97, type: 3}
+  m_Script: {fileID: 11500000, guid: bdaa043e324e04b91a4dcefec1ebb409, type: 3}
   m_Name: DefaultStyle
   m_EditorClassIdentifier: 
   filterStyles:
@@ -46,8 +46,8 @@ MonoBehaviour:
         Width: 15
         MiterLimit: 3
         Enabled: 1
-    - layerName: water
-      Material: {fileID: 2100000, guid: bb0267789bf4345c7916b57b125fc6ca, type: 2}
+    - layerName: roads
+      Material: {fileID: 2100000, guid: 7f88baed2a0f94bb3a42f7bc45b5fcf8, type: 2}
       PolygonBuilderOptions:
         Material: {fileID: 0}
         Extrude: 1
@@ -62,8 +62,8 @@ MonoBehaviour:
         Width: 15
         MiterLimit: 3
         Enabled: 1
-    - layerName: roads
-      Material: {fileID: 2100000, guid: 7f88baed2a0f94bb3a42f7bc45b5fcf8, type: 2}
+    - layerName: water
+      Material: {fileID: 2100000, guid: bb0267789bf4345c7916b57b125fc6ca, type: 2}
       PolygonBuilderOptions:
         Material: {fileID: 0}
         Extrude: 1
@@ -82,5 +82,5 @@ MonoBehaviour:
       CollectionNameSet:
       - buildings
       - earth
-      - water
       - roads
+      - water
