+++ Assets/Scenes/DefaultStyle.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 141951d475bf741588a63ed80ea9396d
-timeCreated: 1504124094
+guid: b85c5ba8553a442a68b2ac5b34f2697c
+timeCreated: 1504293569
 licenseType: Free
 NativeFormatImporter:
   mainObjectFileID: 11400000
